P0 complete: provenance applied to all 367 runs; 2 primitive defects fixed; P10.0 corrects §0.5
PROVENANCE (plan §2.1-2.2): ds_common.write_runmeta/write_done (torch-optional, never
raises) + backfill_runmeta.py + update_registry.py. 734 RUNMETA.json/DONE.json written
across all 367 run dirs, idempotent. 181 dirs recovered a real git commit from logs/*.out,
195 recovered their script from the sbatch wrapper. Every reconstructed field carries
{"source":"reconstructed","evidence":...}; unsourceable fields go to unknown_fields with a
reason - nothing fabricated. EXPERIMENT_REGISTRY.csv 45 -> 395 rows (.bak kept).

TWO PRIMITIVE DEFECTS FOUND BY THE NEW TESTS, BOTH FIXED:
 - pair_common.ComponentCapture._grab: torch.tensor([]) had no dtype -> float32 ->
   index_select raises on an empty position list. Fixed with dtype=torch.long.
 - ds_common.find_word_occurrences_in_text: a stray trailing `ids = enc["input_ids"]`
   raised UnboundLocalError on the slow-tokenizer path AND, when enc was bound but failed
   the sanity check, silently mis-sliced subtoken_ids from a different-length id list.
   The silent mis-slice is worse than the crash the test was written for and was found
   during review. Fixed by deleting the line.
Test suite 113 -> 191 passing (78 new synthetic tests for AttentionKnockout,
ComponentCapture, resolve_positions, SubmodulePatch components), 0 failures, 0 xfail.

VALIDATION: 4909 summary values recomputed from raw across 29 run dirs -> 0 mismatches.
Every behavioral headline number is arithmetically sound. 6/6 negative controls fire.
Found: 1 hard FAIL (the job-708038 aborted twin), 1 §1.5 pooling violation
(phase6_mlpKO_curated_layer_703457 pools dev+heldout), 4 unmarked smoke dirs.

P10.0 CORRECTS PLAN §0.5 (important): the graded-score re-analysis reproduces
+0.0741 (perm p 0.034) BUT FAILS ITS SPECIFICITY CONTROL - the random-head arm gives
+0.0392 (53% of the effect) and the direct contrast rand-minus-carry is +0.0349
[-0.039,+0.110] p=0.382. The targeted ablation is NOT demonstrably better than a
size-matched random one. Fragile too: LOO permutation p reaches 0.0624, curated does not
replicate, BEHAV-WRITE is null on graded as well. The "behaviorally inert" conclusion
STANDS, but the reason is now "indistinguishable from a random ablation at this n"
rather than "we found nothing". Post-hoc power confirmed at 0.135/0.086; n~275 needed.

+++ doublespeak_causality/pair_common.py
@@ -120,8 +120,11 @@ class ComponentCapture:
     def _grab(self, comp, li, hidden):
         if hidden.dim() == 3:
             hidden = hidden[0]
+        # dtype=long is required: torch.tensor([]) defaults to float32, which makes
+        # index_select raise for an empty position list (reachable from
+        # capture_components when every requested position resolves to None).
         idx = torch.tensor([p for p in self.positions if 0 <= p < hidden.shape[0]],
-                           device=hidden.device)
+                           device=hidden.device, dtype=torch.long)
         if idx.numel() != len(self.positions):
             raise IndexError(f"position out of range for seq_len={hidden.shape[0]}")
         self._buf[comp][li] = hidden.index_select(0, idx).float().cpu()
